moved the vmdata to its own module
